Move StartRule and ClauseType to metabase-lib/v1/expressions
