+++ frontend/src/metabase-lib/v1/expressions/complete/complete.ts
@@ -4,12 +4,14 @@ import { isNotNull } from "metabase/lib/types";
 import type * as Lib from "metabase-lib";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
 
+import type { StartRule } from "../types";
+
 export type SuggestOptions = {
   query: Lib.Query;
   stageIndex: number;
   metadata: Metadata;
   reportTimezone?: string;
-  startRule: "expression" | "boolean" | "aggregation";
+  startRule: StartRule;
   expressionIndex: number | undefined;
 };
