+++ frontend/src/metabase-lib/v1/expressions/types.ts
@@ -1,3 +1,4 @@
+import type * as Lib from "metabase-lib";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type { DatabaseFeature, Expression } from "metabase-types/api";
 
@@ -36,6 +37,15 @@ interface HelpTextArg {
   description: string;
   example: Expression | ["args", Expression[]];
 }
+
+type ClauseByStartRule = {
+  expression: Lib.ExpressionClause;
+  aggregation: Lib.AggregationClause;
+  boolean: Lib.FilterClause;
+};
+
+export type StartRule = keyof ClauseByStartRule;
+export type ClauseType<S extends StartRule> = ClauseByStartRule[S];
 
 type MBQLClauseFunctionReturnType =
   | "aggregation"
