+++ frontend/src/metabase/query_builder/components/expressions/Editor/language.ts
@@ -3,11 +3,9 @@ import { type Diagnostic, linter } from "@codemirror/lint";
 import type { EditorView } from "@codemirror/view";
 
 import type * as Lib from "metabase-lib";
-import type { ErrorWithMessage } from "metabase-lib/v1/expressions";
+import type { ErrorWithMessage, StartRule } from "metabase-lib/v1/expressions";
 import { parser } from "metabase-lib/v1/expressions/tokenizer/parser";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
-
-import type { StartRule } from "../types";
 
 import { DEBOUNCE_VALIDATION_MS } from "./constants";
 import { diagnoseAndCompileExpression } from "./utils";
