+++ frontend/src/metabase/query_builder/components/expressions/Editor/utils.ts
@@ -1,13 +1,15 @@
 import { t } from "ttag";
 
 import type * as Lib from "metabase-lib";
-import { isExpression } from "metabase-lib/v1/expressions";
+import {
+  type ClauseType,
+  type StartRule,
+  isExpression,
+} from "metabase-lib/v1/expressions";
 import { diagnoseAndCompile } from "metabase-lib/v1/expressions/diagnostics";
 import { getFunctionByStructure } from "metabase-lib/v1/expressions/helper-text-strings";
 import { parser } from "metabase-lib/v1/expressions/tokenizer/parser";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
-
-import type { ClauseType, StartRule } from "../types";
 
 export function enclosingFunction(doc: string, pos: number) {
   const tree = parser.parse(doc);
